tests/requirements.txt
certifi==2024.6.2
dnspython==2.6.1
httpx==0.27.0
./moddns_client
pytest==8.2.2
pytest-asyncio==0.23.7
pydantic-settings==2.3.4
redis==6.0.0
retry==0.9.2
requests==2.32.3
testcontainers==4.7.0
